Created TranslationUnit class interface.
The current thinking is for this class to hold the representation of the object data (i.e. completely processed object file data) suitable to be written to an ROF file. A parser (which doesn't exist yet) in conjunction with likely a few other classes will construct an instance of this as the input file (asm, .a) is read. The instance will then be passed on to a writer class which will format its data into and ROF file.