Restructure project to form OS9 Toolchain.
It seems likely that many OS9 tools (e.g. 'ident' and 'amips') will end up sharing the same data structures and algorithms. This commit merges the work I'd already done on implementing an OS9 module info dump tool (now named 'ident' in accordance with the name of the official tool), pulling out common classes into a working tree hierarchy which aims to be similar to that of the LLVM project.

Other changes:
- Renames the MIPS assembler project/tool to 'amips' (my best guess as to what the official OS9 toolchain would have called this assembler)..
- Adds an empty project 'bemips', which would contain a tool used to translate ICode (Ultra C++'s bitcode equivalent in LLVM speak) into MIPS ROF object files. (The name 'bemips' is also just a guess).