[amips][Assembler] Move requires arg reading check to compile-time.
This makes use of "if constexpr" within expression parsers (RType(), IType(), etc.) to conditionally compile in reading of arguments from the operands field. This works by introducing a sentinel (called "Arg") which, when provided as the numeric constant expression value for a parser template argument, is recognized by the parser function's corresponding "constexpr if" check. If the sentinel is provided, the corresponding operand will be read from the operands parser.

For example:

  IType<0b101010, Arg, 0b11111, Arg, ITypeTuple<RS, Immediate>>(...);

would handle an instruction that looks like:

  mnem $at,0x555

The first "Arg" in the template parameter list generates code to read register RS from the operands list ($at). The second "Arg" generates code to read the Immediate from the operands list (0x555).

Previously, without the sentinel, we had to check if an argument was provided for every kind of argument for the corresponding instruction format type, even if the instruction we were defining could never semantically receive that argument as a user-specified value.