[test][Assembler] Move equality and print functions to common files.
I considered moving at least equality check functions to the Assembler library itself, but today, only the test library has a usecase. The only existing == function defined is for Expression types, and its current implementation  relies on rtti. Perhaps if it were rewritten using a visitor (e.g. the expression printer), it'd be a better candidate to include within the Assembler library itself.