[amips][Assembler][test] Add unit testing for MipsAssemblerTarget.
This change introduces the Capstone disassembler library as a dependency (managed via CMake via PkgConfig), which is used to ensure that the instructions emitted by MipsAssemblerTarget are correct. This is checked by first encoding a human readable instruction in OS9 asm syntax, decoding the result with Capstone, and then asserting that the original input text equates to what Capstone spits out.