[Assembler] Begin implementing pseudo instructions "align" and "ds".
In order to support data operations like these, we need to support tracking of data counters as well as the tracking of labels assigned to their locations.

I've begun to do this here, but this code will likely change significantly as it becomes more obvious how these things should be handled cleanly.