[Support] Add Numeric helpers.
Currently this includes a helper struct (template metaprogramming) which provides the maximum value that a numeric type can hold, stripping cv and ref with std::decay.

This may not be a best practice, since it's not obvious from the client's perspective that cvref will be stripped, but it makes client code cleaner. Perhaps it's not so bad, since a client who manually strips such qualifiers would still get the same behavior with this implementation.