Create abstract class and implementations for Assembler operation handlers.
This way, handler classes can be registered in a particular order with the assembler, and the first one registered that is capable of handling the given operation will do so.

Previously, we had a HandleDirective and a HandlePseudoInst method that we'd call to do this same thing. Now, it'd be easy to register additional handlers to extend the assembler's functionality (or to encapsulate a CPU target as just another handler in the future).

A primary motivator of this design is to make individual handlers easily unit testable.