[WIP] Add universal ObjectFile representation.
This commit plays around with project structure a bit, and moves Expression and ObjectFile to their own namespaces.

I wrote this commit message in retrospect (it was originally just a reminder to squash it with something else), indicative of a period
of massive refactoring, some good, some bad, as a result of taking time away from the project and forgetting where I was when I'd left off.