Add CI/CD pipeline.
Currently, we just build and test on macOS, because brew has the necessary version of capstone (4.0.1),
whereas Ubuntu 18.04 does not. Ultimately, we should run CI with Linux since the GitHub runner uses the
least amount of monthly 'minutes'.

In the interim, this should be fine, since we only trigger CI runs on a merge to master, and most work
will happen in feature branches.