Fix undefined behavior in MipsAssemblerTarget when accessing operands list.
This fixes an issue with single argument JALR instruction, which depends on an out-of-range access exception to determine
that the JALR should default its second operand.

It also prevents undefined behavior when an expected operand is missing. However, this is still not properly handled, in that
a proper error should be propagated to the user when an operand is incorrect. For consistency, we should reuse Operation.h for
targets, which has this functionality built in.