Partially implement expression tree reference serialization support.
Todo:
- set and equ expression tree refs
- alignment requirement specification for expression tree external references.

Regarding alignment requirements, I *think* they would be used to inform the linker of the alignment requirement of the *context*
in which a reference appears. For example, if a reference is the target of a jump instruction, that reference would likely need
to be 4-byte aligned. But, if this is the case, why wouldn't this be done instead inside the ROF's reference data section?
If they were, we would only need to care that the value of the fully resolved expression tree after linking met the alignment
requirement. Perhaps this would be too wishy-washy, or maybe it tells us that the linker will actually attempt to lay out
all ROFs in a way that satisfies all alignment requirements.

Further, why wouldn't alignment requirements be applicable for local references (they aren't, according to docs)?
That seems like another hint that perhaps the linker will actively attempt to acheive the alignment requirement
when linking (since a single ROF is not a candidate for layout changes, perhaps the linker doesn't care about local alignment).